Inline class; extracted too early.
