+++ .idea/modules/ispd-plugin.iml
@@ -1,0 +1,6 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<module version="4">
+  <component name="SonarLintModuleSettings">
+    <option name="uniqueId" value="a2efa86c-962f-4f05-a372-f375cbbb4c1a" />
+  </component>
+</module>
