+++ .idea/modules/ispd-plugin.main.iml
@@ -1,0 +1,6 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<module version="4">
+  <component name="SonarLintModuleSettings">
+    <option name="uniqueId" value="d9f5ac99-414d-4e53-8a29-628fed6ce906" />
+  </component>
+</module>
