+++ .idea/modules/ispd-plugin.test.iml
@@ -1,0 +1,6 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<module version="4">
+  <component name="SonarLintModuleSettings">
+    <option name="uniqueId" value="a81cd015-c951-487e-8180-ff89fb0cced8" />
+  </component>
+</module>
